[#154747783] Split files ListingDetailsContentCard
